[PM-6702] Move password strength indicator from createaccount to completeregistration
